Refactor : test
테스트 코드 우선순위 낮추기
