--- server/src/test/java/toyspringboot/server/JWT/JwtTokenProviderTest.java
@@ -1,11 +1,0 @@
-//package toyspringboot.server.JWT;
-//
-//import org.junit.jupiter.api.DisplayName;
-//import org.junit.jupiter.api.Test;
-//import org.springframework.beans.factory.annotation.Autowired;
-//import toyspringboot.server.ServerApplicationTests;
-//
-//public class JwtTokenProviderTest extends ServerApplicationTests {
-//    @Autowired
-//    private TokenProvider tokenProvider;
-//}
