--- server/src/test/java/toyspringboot/server/Notice/NoticeFailConstants.java
@@ -1,9 +1,0 @@
-package toyspringboot.server.Notice;
-
-public class NoticeFailConstants {
-    // ---------- API ----------
-
-    // ---------- test information ----------
-
-    // ---------- response message ----------
-}
