--- server/src/test/java/toyspringboot/server/TestConfig/DomainTest.java
@@ -1,15 +1,0 @@
-package toyspringboot.server.TestConfig;
-
-
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
-import toyspringboot.server.ServerApplicationTests;
-
-
-@DataJpaTest
-@Transactional(propagation = Propagation.NOT_SUPPORTED)
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-public class DomainTest extends ServerApplicationTests {
-}
