--- server/src/test/java/toyspringboot/server/TestConfig/ServiceTest.java
@@ -1,9 +1,0 @@
-package toyspringboot.server.TestConfig;
-
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
-import toyspringboot.server.ServerApplicationTests;
-
-@ExtendWith(MockitoExtension.class)
-public class ServiceTest extends ServerApplicationTests {
-}
